From Victor Update
CVss score can be exported. At present, only OWASP top ten 2017 information (a1-a10) is poor. But I have a look at the meaning of Ian Ashworth. He means that there is no way to get this value directly. You can only write down the mapping relationship between CWE and OWASP in the program, and map this value through CWE. I don't know that I understand it correctly. If so, he needs to Send a piece of code to have a look. I don't know what the mapping relationship is.